+++ cmd/sidecars/kafka_new/client/environment.go
@@ -1,0 +1,51 @@
+package kafkasc
+
+import (
+	"os"
+
+	"go.uber.org/zap"
+)
+
+// Environment represents the current inspr environment
+type Environment struct {
+	KafkaBootstrapServers string
+	KafkaAutoOffsetReset  string
+}
+
+var env *Environment
+var logger *zap.Logger
+
+// init is called after all the variable declarations in the package have evaluated
+// their initializers, and those are evaluated only after all the imported packages
+// have been initialized
+func init() {
+	logger, _ = zap.NewDevelopment(zap.Fields(zap.String("section", "kafka-sidecar")))
+}
+
+// GetEnvironment returns the current inspr environment
+func GetEnvironment() *Environment {
+	if env == nil {
+		env = &Environment{
+			KafkaBootstrapServers: getEnv("KAFKA_BOOTSTRAP_SERVERS"),
+			KafkaAutoOffsetReset:  getEnv("KAFKA_AUTO_OFFSET_RESET"),
+		}
+	}
+	return env
+}
+
+func getEnv(name string) string {
+	if value, exists := os.LookupEnv(name); exists {
+		return value
+	}
+	panic("[ENV VAR] " + name + " not found")
+}
+
+// RefreshEnviromentVariables "refreshes" the value of kafka environment variables.
+// This was develop for testing and probably sholdn't be used in other cases.
+func RefreshEnviromentVariables() *Environment {
+	env = &Environment{
+		KafkaBootstrapServers: getEnv("KAFKA_BOOTSTRAP_SERVERS"),
+		KafkaAutoOffsetReset:  getEnv("KAFKA_AUTO_OFFSET_RESET"),
+	}
+	return env
+}