+++ pkg/environment/utils.go
@@ -5,28 +5,33 @@ import (
 	"strings"
 
 	"github.com/inspr/inspr/pkg/ierrors"
+	"github.com/inspr/inspr/pkg/meta/brokers"
 	"github.com/inspr/inspr/pkg/utils"
 )
 
 // IsInChannelBoundary - checks if given channel exists in given boundary
-func IsInChannelBoundary(channel, boundary string) bool {
+func IsInChannelBoundary(channel string, boundary []brokers.ChannelBroker) bool {
 	channelsList := GetChannelBoundaryList(boundary)
 	return utils.Includes(channelsList, channel)
 }
 
 // GetChannelBoundaryList returns a string slice with the channels in given boundary
-func GetChannelBoundaryList(boundary string) utils.StringArray {
-	if boundary == "" {
+func GetChannelBoundaryList(channels []brokers.ChannelBroker) utils.StringArray {
+	if channels == nil {
 		return utils.StringArray{}
 	}
-	return strings.Split(boundary, ";")
+	boundary := utils.StringArray{}
+	for _, channel := range channels {
+		boundary = append(boundary, channel.ChName)
+	}
+	return boundary
 }
 
 // GetResolvedBoundaryChannelList gets the list of resolved channels from given boundary
-func GetResolvedBoundaryChannelList(boundary string) utils.StringArray {
-	channels := utils.StringArray(GetChannelBoundaryList(boundary))
-	return channels.Map(func(s string) string {
-		resolved, _ := GetResolvedChannel(s, boundary, "")
+func GetResolvedBoundaryChannelList(channels []brokers.ChannelBroker) utils.StringArray {
+	boundary := GetChannelBoundaryList(channels)
+	return boundary.Map(func(s string) string {
+		resolved, _ := GetResolvedChannel(s, channels, nil)
 		return resolved
 	})
 }
@@ -42,16 +47,26 @@ func GetSchema(channel string) (string, error) {
 
 // OutputChannnelList returns a list of input channels
 func OutputChannnelList() utils.StringArray {
-	return GetChannelBoundaryList(GetOutputChannels())
+	return GetChannelBoundaryList(GetOutputChannelsData())
 }
 
 // InputChannelList returns a list of input channels
 func InputChannelList() utils.StringArray {
-	return GetChannelBoundaryList(GetInputChannels())
+	return GetChannelBoundaryList(GetInputChannelsData())
+}
+
+// OutputBrokerChannnels returns a list of input channels
+func OutputBrokerChannnels(broker string) utils.StringArray {
+	return GetChannelBoundaryList(GetOutputChannelsData()) // WRONG IMPLEMENTATION
+}
+
+// InputBrokerChannels returns a list of input channels
+func InputBrokerChannels(broker string) utils.StringArray {
+	return GetChannelBoundaryList(GetInputChannelsData()) // WRONG IMPLEMENTATION
 }
 
 // GetResolvedChannel gets a resolved channel from a channel name
-func GetResolvedChannel(channel, inputChan, outputChan string) (string, error) {
+func GetResolvedChannel(channel string, inputChan, outputChan []brokers.ChannelBroker) (string, error) {
 	if IsInChannelBoundary(channel, inputChan) || IsInChannelBoundary(channel, outputChan) {
 		return os.Getenv(channel + "_RESOLVED"), nil
 	}
@@ -59,4 +74,32 @@ func GetResolvedChannel(channel, inputChan, outputChan string) (string, error) {
 		InvalidChannel().
 		Message("channel " + channel + " not listed as an input or output").
 		Build()
+}
+
+func getChannelData(channelList string) []brokers.ChannelBroker {
+	if channelList == "" {
+		return nil
+	}
+	channelBrokers := utils.StringArray(strings.Split(channelList, ";"))
+	channels := []brokers.ChannelBroker{}
+	channelBrokers.Map(func(channel string) string {
+		data := strings.Split(channel, "_")
+		channelData := brokers.ChannelBroker{
+			ChName: data[0],
+			Broker: data[1],
+		}
+		channels = append(channels, channelData)
+		return ""
+	})
+	return channels
+}
+
+func filterChannelsByBroker(broker string, channels []brokers.ChannelBroker) utils.StringArray {
+	brokerChannels := utils.StringArray{}
+	for _, channel := range channels {
+		if channel.Broker == broker {
+			brokerChannels = append(brokerChannels, channel.ChName)
+		}
+	}
+	return brokerChannels
 }