+++ pkg/sidecars/models/interfaces.go
@@ -1,14 +1,18 @@
 package models
 
-import "context"
+import (
+	"context"
+)
 
 // Reader reads from a message broker
 type Reader interface {
-	ReadMessage(ctx context.Context, channel string) (BrokerMessage, error)
+	ReadMessage(ctx context.Context, channel string) ([]byte, error)
 	Commit(ctx context.Context, channel string) error
+	Close() error
 }
 
 // Writer writes messages in a message broker
 type Writer interface {
-	WriteMessage(channel string, msg interface{}) error
+	WriteMessage(channel string, msg []byte) error
+	Close()
 }