+++ pkg/sidecars/models/types.go
@@ -4,3 +4,10 @@ package models
 type BrokerMessage struct {
 	Data interface{} `json:"data"`
 }
+
+// ConnectionVariables is the structure resposible for storing
+// enviroment variable names regarding connection ports for sidecars
+type ConnectionVariables struct {
+	ReadEnvVar  string
+	WriteEnvVar string
+}