+++ pkg/sidecars/server/handlers_test.go
@@ -1,9 +1,11 @@
 package sidecarserv
 
 import (
+	"bytes"
 	"context"
 	"encoding/json"
 	"errors"
+	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"os"
@@ -12,9 +14,10 @@ import (
 	"testing"
 	"time"
 
+	"github.com/confluentinc/confluent-kafka-go/kafka"
+	kafkasc "github.com/inspr/inspr/cmd/sidecars/kafka_new/client"
 	"github.com/inspr/inspr/pkg/rest"
-	"github.com/inspr/inspr/pkg/rest/request"
-	"github.com/inspr/inspr/pkg/sidecar_old/models"
+	"github.com/inspr/inspr/pkg/sidecars/models"
 )
 
 // createMockEnvVars - sets up the env values to be used in the tests functions
@@ -24,6 +27,8 @@ func createMockEnvVars() {
 	os.Setenv("INSPR_INPUT_CHANNELS", customEnvValues)
 	os.Setenv("INSPR_OUTPUT_CHANNELS", customEnvValues)
 	os.Setenv("INSPR_UNIX_SOCKET", unixSocketAddr)
+	os.Setenv("INSPR_LBSIDECAR_READ_PORT", "8020")
+	os.Setenv("INSPR_SIDECAR_KAFKA_WRITE_PORT", "8021")
 	os.Setenv("INSPR_APP_CTX", "random.ctx")
 	os.Setenv("INSPR_ENV", "test")
 	os.Setenv("INSPR_APP_ID", "appid")
@@ -31,75 +36,79 @@ func createMockEnvVars() {
 
 // deleteMockEnvVars - deletes the env values used in the tests functions
 func deleteMockEnvVars() {
-	os.Unsetenv("INSPR_OUTPUT_CHANNELS")
 	os.Unsetenv("INSPR_INPUT_CHANNELS")
+	os.Unsetenv("INSPR_OUTPUT_CHANNELS")
 	os.Unsetenv("INSPR_UNIX_SOCKET")
+	os.Unsetenv("INSPR_LBSIDECAR_READ_PORT")
+	os.Unsetenv("INSPR_SIDECAR_KAFKA_WRITE_PORT")
 	os.Unsetenv("INSPR_APP_CTX")
 	os.Unsetenv("INSPR_ENV")
 	os.Unsetenv("INSPR_APP_ID")
 }
 
 type mockReader struct {
-	readMessage func(ctx context.Context, channel string) (models.BrokerData, error)
+	readMessage func(ctx context.Context, channel string) ([]byte, error)
 	commit      func(ctx context.Context, channel string) error
 }
 
-func (m mockReader) Commit(ctx context.Context, channel string) error {
+func (m *mockReader) Commit(ctx context.Context, channel string) error {
 	return m.commit(ctx, channel)
 }
-func (m mockReader) ReadMessage(ctx context.Context, channel string) (models.BrokerData, error) {
+func (m *mockReader) ReadMessage(ctx context.Context, channel string) ([]byte, error) {
 	return m.readMessage(ctx, channel)
 }
 
+func (m *mockReader) Close() error { return nil }
+
+func (m *mockReader) Consumers() map[string]kafkasc.Consumer { return nil }
+
 type mockWriter struct {
-	writeMessage func(channel string, message interface{}) error
+	writeMessage func(channel string, message []byte) error
 }
 
-func (m mockWriter) WriteMessage(channel string, message interface{}) error {
+func (m *mockWriter) WriteMessage(channel string, message []byte) error {
 	return m.writeMessage(channel, message)
 }
+
+func (m *mockWriter) Close() {}
+
+func (m *mockWriter) Producer() *kafka.Producer { return nil }
 func TestServer_writeMessageHandler(t *testing.T) {
 	createMockEnvVars()
 	defer deleteMockEnvVars()
 	tests := []struct {
-		readerFunc func(t *testing.T) mockReader
-		writerFunc func(t *testing.T) mockWriter
+		readerFunc func(t *testing.T) models.Reader
+		writerFunc func(t *testing.T) models.Writer
 		channel    string
 		name       string
 		wantErr    bool
-		message    interface{}
+		message    []byte
 	}{
 		{
 			name:    "correct behaviour test",
 			channel: "chan",
-			message: struct {
-				Message interface{} `json:"message"`
-			}{"lofi nordeste"},
+			message: []byte("lofi nordeste"),
 		},
 		{
 			name:    "invalid channel",
 			channel: "invalid",
-			message: struct {
-				Message interface{} `json:"message"`
-			}{"lofi nordeste"},
+			message: []byte("lofi nordeste"),
 			wantErr: true,
 		},
 		{
 			name:    "invalid data for marshalling",
 			channel: "chan",
-			message: "invalid message",
+			message: []byte("invalid message"),
 			wantErr: true,
 		},
 		{
 			name:    "invalid broker response",
 			channel: "chan",
-			message: struct {
-				Message interface{} `json:"message"`
-			}{"this is an invalid message"},
+			message: []byte("this is an invalid message"),
 			wantErr: true,
-			writerFunc: func(t *testing.T) mockWriter {
-				return mockWriter{
-					writeMessage: func(channel string, message interface{}) error {
+			writerFunc: func(t *testing.T) models.Writer {
+				return &mockWriter{
+					writeMessage: func(channel string, message []byte) error {
 						return errors.New("this is an error")
 					},
 				}
@@ -109,14 +118,11 @@ func TestServer_writeMessageHandler(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			if tt.writerFunc == nil {
-				tt.writerFunc = func(t *testing.T) mockWriter {
-					return mockWriter{
-						writeMessage: func(channel string, message interface{}) error {
-							if !reflect.DeepEqual(message, tt.message.(struct {
-								Message interface{} `json:"message"`
-							}).Message) {
-
-								t.Errorf("Server_writeMessageHandler message = %v, want = %v", message, tt.message)
+				tt.writerFunc = func(t *testing.T) models.Writer {
+					return &mockWriter{
+						writeMessage: func(channel string, message []byte) error {
+							if !reflect.DeepEqual(message, tt.message) {
+								t.Errorf("Server_writeMessageHandler message = %v, want = %v", string(message), string(tt.message))
 							}
 
 							return nil
@@ -129,17 +135,15 @@ func TestServer_writeMessageHandler(t *testing.T) {
 			}
 			server := httptest.NewServer(s.writeMessageHandler())
 			defer server.Close()
-			client := request.NewJSONClient(server.URL)
+			client := &http.Client{}
 
-			rest := struct {
-				Status string `json:"status"`
-			}{}
-
-			err := client.Send(context.Background(), tt.channel, http.MethodPost, tt.message, &rest)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Server_writeMessageHandler err = %v, wantErr = %v", err, tt.wantErr)
+			resp, err := client.Post(fmt.Sprintf("%s/%s", server.URL, tt.channel), "application/octet-stream", bytes.NewBuffer(tt.message))
+			if err != nil || resp.StatusCode != http.StatusOK {
+				err = rest.UnmarshalERROR(resp.Body)
+				if (err != nil) != tt.wantErr {
+					t.Errorf("Server_writeMessageHandler err = %v, wantErr = %v", err, tt.wantErr)
+				}
 			}
-
 		})
 	}
 }
@@ -147,31 +151,29 @@ func TestServer_writeMessageHandler(t *testing.T) {
 func TestServer_readMessageRoutine(t *testing.T) {
 	createMockEnvVars()
 	defer deleteMockEnvVars()
-	channels := map[string]chan interface{}{
-		"chan":    make(chan interface{}, 2),
-		"banana":  make(chan interface{}, 2),
-		"testing": make(chan interface{}, 2),
+	channels := map[string]chan []byte{
+		"chan":    make(chan []byte, 2),
+		"banana":  make(chan []byte, 2),
+		"testing": make(chan []byte, 2),
 	}
-	readerFuncErr := func(t *testing.T) mockReader {
-		return mockReader{
-			readMessage: func(ctx context.Context, channel string) (models.BrokerData, error) {
-				return models.BrokerData{}, errors.New("this is an error")
+	readerFuncErr := func(t *testing.T) models.Reader {
+		return &mockReader{
+			readMessage: func(ctx context.Context, channel string) ([]byte, error) {
+				return nil, errors.New("this is an error")
 			},
 		}
 	}
-	readerFunc := func(t *testing.T) mockReader {
-		return mockReader{
-			readMessage: func(ctx context.Context, channel string) (models.BrokerData, error) {
-				var msg interface{}
+	readerFunc := func(t *testing.T) models.Reader {
+		return &mockReader{
+			readMessage: func(ctx context.Context, channel string) ([]byte, error) {
+				var msg []byte
 				select {
 
 				case msg = <-channels[channel]:
 				case <-ctx.Done():
-					return models.BrokerData{}, ctx.Err()
+					return nil, ctx.Err()
 				}
-				return models.BrokerData{
-					Message: msg,
-				}, nil
+				return msg, nil
 			},
 			commit: func(ctx context.Context, channel string) error {
 				return nil
@@ -179,12 +181,12 @@ func TestServer_readMessageRoutine(t *testing.T) {
 		}
 	}
 	tests := []struct {
-		readerFunc func(t *testing.T) mockReader
-		writerFunc func(t *testing.T) mockWriter
+		readerFunc func(t *testing.T) models.Reader
+		writerFunc func(t *testing.T) models.Writer
 		channel    string
 		name       string
 		wantErr    bool
-		message    interface{}
+		message    []byte
 	}{
 
 		{
@@ -245,7 +247,9 @@ func TestServer_readMessageRoutine(t *testing.T) {
 			))
 			defer server.Close()
 
-			s.client = request.NewJSONClient(server.URL)
+			s.outAddr = server.URL
+			s.client = &http.Client{}
+			s.broker = "someBroker"
 			ctx, cancel := context.WithTimeout(context.Background(), time.Millisecond*300)
 			defer cancel()
 			go s.readMessageRoutine(ctx)
@@ -257,6 +261,8 @@ func TestServer_readMessageRoutine(t *testing.T) {
 				if received == tt.wantErr {
 					t.Errorf("Server_readMessageRoutine received = %v, wantErr = %v", received, tt.wantErr)
 				}
+			default:
+				return
 			}
 		})
 	}