+++ pkg/utils/string_slice.go
@@ -12,10 +12,8 @@ import (
 	kubeCore "k8s.io/api/core/v1"
 )
 
-/*
-Index returns the first index of the target string t,
-or -1 if no match is found.
-*/
+// Index returns the first index of the target string t,
+// or -1 if no match is found.
 func Index(vs []string, t string) int {
 	for i, v := range vs {
 		if v == t {
@@ -25,17 +23,13 @@ func Index(vs []string, t string) int {
 	return -1
 }
 
-/*
-Includes returns true if the target string t is in the slice.
-*/
+// Includes returns true if the target string t is in the slice.
 func Includes(vs []string, t string) bool {
 	return Index(vs, t) >= 0
 }
 
-/*
-Remove return a new slice without any occurrence of the
-target string t
-*/
+// Remove return a new slice without any occurrence of the
+// target string t
 func Remove(vs []string, t string) []string {
 	var newSlice []string
 	for _, v := range vs {
@@ -46,10 +40,8 @@ func Remove(vs []string, t string) []string {
 	return newSlice
 }
 
-/*
-StringSliceUnion returns the sorted union of the two string slices.
-Remember that union dont have repeated elements
-*/
+// StringSliceUnion returns the sorted union of the two string slices.
+// Remember that union dont have repeated elements
 func StringSliceUnion(a, b []string) []string {
 	check := make(map[string]bool)
 
@@ -66,9 +58,7 @@ func StringSliceUnion(a, b []string) []string {
 	return res
 }
 
-/*
-Map returns a new slice containing the results of applying the function f to each string in the original slice.
-*/
+// Map returns a new slice containing the results of applying the function f to each string in the original slice.
 func Map(vs []string, f func(string) string) []string {
 	vsm := make([]string, len(vs))
 	for i, v := range vs {